requirements.txt
# --- API / App ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# --- Frontend ---
streamlit==1.41.1
requests==2.32.3

# --- LLM / Embeddings / Orchestration ---
ollama==0.4.4
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-text-splitters==0.3.4
langgraph==0.2.60

# --- Vector DB & Retrieval ---
chromadb==0.5.23
rank-bm25==0.2.2

# --- NLP ---
spacy==3.8.3
langdetect==1.0.9

# --- Document parsing ---
pypdf==5.1.0
python-docx==1.1.2
Pillow==11.0.0
pytesseract==0.3.13
pdf2image==1.17.0

# --- Logging & Observability ---
loguru==0.7.3

# --- Testing ---
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.27.2

# --- Numerics ---
numpy==1.26.4
